sessions: consolidate new-session/new-chat pill view items
Move CompactNewButtonActionViewItem, NewSessionActionViewItem and
NewSessionActionViewItemContribution from newSessionActionViewItem.ts into
sessionsActions.ts (next to NewChatActionViewItem) and delete the now-empty
file. Update sessions.contribution.ts to import the contribution from
sessionsActions.js.
